% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{explore.sg}
\alias{explore.sg}
\title{explore.sg}
\usage{
explore.sg(
  x.orig,
  y.orig,
  xmin,
  xmax,
  ymin,
  ymax,
  w0 = 3,
  w1 = 5,
  w2 = 7,
  p0 = 0,
  p1 = 1,
  p2 = 2,
  tag = "tag"
)
}
\arguments{
\item{tag}{}
}
\value{

}
\description{
explore.sg
}
