% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.query.x.open.end}
\alias{get.query.x.open.end}
\title{get.query.x.open.end}
\usage{
get.query.x.open.end(
  ref,
  query,
  lm.ref,
  lm.query,
  ngrid = 100,
  scaling = "linear"
)
}
\arguments{
\item{scaling}{}
}
\value{

}
\description{
get.query.x.open.end
}
