% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vachette_functions.R
\name{get.ref.x.open.end}
\alias{get.ref.x.open.end}
\title{Determine last x value for open end reference curve}
\usage{
get.ref.x.open.end(x, y, lm, step.x.factor = 1.5, tol = 0.01)
}
\arguments{
\item{x}{x values}

\item{y}{y values}

\item{lm}{landmarks}

\item{step.x.factor}{factor to increase step.x}

\item{tol}{tolerance for max and min identification}
}
\description{
Determine last x value for open end reference curve
}
