% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vachette_functions.R
\name{get.x.multi.landmarks}
\alias{get.x.multi.landmarks}
\title{Get x multi landmarks}
\usage{
get.x.multi.landmarks(x, y, w = 17, tol = 1e-09)
}
\arguments{
\item{x}{x values}

\item{y}{y values}

\item{w}{window size}

\item{tol}{tolerance for max and min identification}
}
\description{
Find initial landmark position using f1 and f2 derivatives and Savitzky Golay smoothing
Default window = 17 gridpoints
}
