% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.y.diff.ref.query}
\alias{get.y.diff.ref.query}
\title{get.y.diff.ref.query}
\usage{
get.y.diff.ref.query(
  x,
  query.x.start,
  query,
  refGrid,
  ngrid = 100,
  scaling = "linear"
)
}
\arguments{
\item{scaling}{}
}
\value{

}
\description{
With constant or linearly changing scaling factor, with some default values
}
