% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{multi.approx}
\alias{multi.approx}
\title{multi.approx}
\usage{
multi.approx(x, y, yout, tol = 1e-09)
}
\arguments{
\item{tol}{}
}
\value{
Return multiple x hits of (x,y) series for y-value yout and returns NA if no hit
}
\description{
multi.approx
}
