% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{multi.peaks}
\alias{multi.peaks}
\title{multi.peaks}
\usage{
multi.peaks(x, y)
}
\arguments{
\item{y}{}
}
\value{
Return multiple x hits of (x,y) series for y-value yout and returns NA if no hit
}
\description{
multi.peaks
}
