% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oral_absorption_sim}
\alias{oral_absorption_sim}
\title{Oral Absorption Simulated Data}
\format{
A data frame with the following columns:
\describe{
  \item{\code{REP}}{Replicate number}
  \item{\code{ID}}{Subject ID}
  \item{\code{time}}{Time used as x-variable}
  \item{\code{EVID}}{Dose event column}
  \item{\code{DV}}{Dependent variable}
  \item{\code{PRED}}{Population predicted value}
  \item{\code{WT}}{Weight covariate}
}
}
\usage{
oral_absorption_sim
}
\description{
Simulated dataset containing 100 replicates. Time points are
simulated given observation time points in  \code{\link{oral_absorption_obs}}.
}
\examples{
\dontrun{
data(oral_absorption_sim)
head(oral_absorption_sim)
}
}
\keyword{datasets}
