% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oral_absorption_typ}
\alias{oral_absorption_typ}
\title{Oral Absorption Typical Curves Data}
\format{
A data frame with the following columns:
\describe{
  \item{\code{ID}}{Subject ID}
  \item{\code{time}}{Time used as x-variable}
  \item{\code{EVID}}{Dose event column}
  \item{\code{PRED}}{Population predicted value}
  \item{\code{WT}}{Weight covariate}
}
}
\usage{
oral_absorption_typ
}
\description{
This dataset contains a single covariate \code{WT} and a single dose.
The dose event row is recorded by the \code{EVID} column.
}
\examples{
\dontrun{
data(oral_absorption_typ)
head(oral_absorption_typ)
}
}
\keyword{datasets}
