% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{p.obs.ref.query}
\alias{p.obs.ref.query}
\title{p.obs.ref.query}
\usage{
p.obs.ref.query(vachette_data)
}
\arguments{
\item{vachette_data}{Object of class \code{vachette_data}}
}
\value{
Object of class \code{ggplot2}
}
\description{
p.obs.ref.query
}
