% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{p.scaled.typical.curves}
\alias{p.scaled.typical.curves}
\title{p.scaled.typical.curves}
\usage{
p.scaled.typical.curves(vachette_data, log.x = FALSE)
}
\arguments{
\item{vachette_data}{Object of class \code{vachette_data}}

\item{log.x}{Set to \code{TRUE} to plot x axis on log scale}
}
\value{
Object of class \code{ggplot2}
}
\description{
p.scaled.typical.curves
}
