% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{p.scaled.typical.full.curves.landmarks}
\alias{p.scaled.typical.full.curves.landmarks}
\title{p.scaled.typical.full.curves.landmarks}
\usage{
p.scaled.typical.full.curves.landmarks(vachette_data, log.x = FALSE)
}
\arguments{
\item{vachette_data}{Object of class\code{vachette_data}}

\item{log.x}{Set to \code{TRUE} to plot x axis on log scale}
}
\description{
p.scaled.typical.full.curves.landmarks
}
