% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{p.transformation}
\alias{p.transformation}
\title{p.transformation}
\usage{
p.transformation(vachette_data)
}
\arguments{
\item{vachette_data}{Object of class \code{vachette_data}}
}
\value{
Object of class \code{ggplot2}
}
\description{
p.transformation
}
