% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vachette_functions.R
\name{print.vachette_data}
\alias{print.vachette_data}
\title{Print \code{vachette_data}}
\usage{
\method{print}{vachette_data}(x, ...)
}
\arguments{
\item{x}{\code{vachette_data} object.}

\item{...}{additional args.}
}
\value{
\code{x} invisibly.
}
\description{
Print generic used to return information about \code{vachette_data} object
}
