% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{refine.x.multi.landmarks}
\alias{refine.x.multi.landmarks}
\title{refine.x.multi.landmarks}
\usage{
refine.x.multi.landmarks(x, y, lm, tol = 1e-09, w1 = 5, w2 = 7)
}
\arguments{
\item{w2}{}
}
\value{

}
\description{
refine.x.multi.landmarks
}
