% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{summary.vachette_data}
\alias{summary.vachette_data}
\title{Summarise \code{vachette_data}}
\usage{
\method{summary}{vachette_data}(object, verbose = FALSE, trim = TRUE, log_file = NULL, ...)
}
\arguments{
\item{object}{\code{vachette_data} object.}

\item{verbose}{logical; Set to \code{TRUE} to view additional warnings output}

\item{trim}{logical; If \code{TRUE}, only the first 20 rows are printed for \code{data.frame} output. Set to \code{FALSE} to display all rows}

\item{log_file}{character; File path to direct console output e.g., \code{"log.txt"}}

\item{...}{additional args.}
}
\value{
\code{x} invisibly.
}
\description{
Summary generic used to return information about \code{vachette_data} object
}
