% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{theme_vachette}
\alias{theme_vachette}
\title{Plot theme for vachette}
\usage{
theme_vachette()
}
\value{
\code{theme, gg}
}
\description{
A ggplot2 theme for vachette plot output
}
\examples{
theme_vachette()

}
