% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vachette_data.R
\name{vachette_data}
\alias{vachette_data}
\title{Initialize vachette object with required data}
\usage{
vachette_data(
  indivsam.obs,
  output.typ,
  indivsam.vpc = NULL,
  vachette.covs,
  ref.dose,
  IIV_CORR = FALSE,
  error = c("proportional", "additive"),
  model.name = NULL,
  mappings = NULL
)
}
\arguments{
\item{indivsam.obs}{Observed data of class \code{data.frame}}

\item{output.typ}{Typ data of class \code{data.frame}}

\item{indivsam.vpc}{VPC data of class  \code{data.frame}}

\item{vachette.covs}{Named character vector of covariate names and reference values}

\item{ref.dose}{Reference dose value in data}

\item{IIV_CORR}{Logical; if IIV corrections performed}

\item{error}{Character; options are \code{"proportional"} or \code{"additive"}. Default is \code{"proportional"}}

\item{model.name}{Character; optional model name for plot output}

\item{mappings}{Named character vector specifying model variable name to column name in input data}
}
\value{
`vachette_data`
}
\description{
Initialize vachette object with required data
}
\details{
Required column names in input \code{data.frame} are:
\itemize{
 \item{"ID"}{ - Subject ID}
 \item{"x"}{ - Typically time}
 \item{"PRED"}{ - Individual prediction}
 \item{"IPRED"}{ - Population prediction}
 \item{"OBS"}{ - DV}
}
If the above column names are different in your input data, use the \code{mappings} argument.
}
\examples{
indivsam.obs <- read.csv(system.file(package = "vachette", "examples", "iv-obs.csv"))
output.typ <- read.csv(system.file(package = "vachette", "examples", "iv-typ.csv"))
indivsam.vpc <- read.csv(system.file(package = "vachette", "examples", "iv-vpc.csv"))


}
