% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vachette_data.R
\name{apply_transformations}
\alias{apply_transformations}
\alias{apply_transformations.vachette_data}
\title{Apply vachette transformations}
\usage{
apply_transformations(vachette_data, ...)

\method{apply_transformations}{vachette_data}(
  vachette_data,
  lm.refine = FALSE,
  tol.end = 0.001,
  tol.noise = 1e-08,
  step.x.factor = 1.5,
  ngrid.fit = 100,
  window = 17,
  window.d1.refine = 7,
  window.d2.refine = 5
)
}
\arguments{
\item{vachette_data}{object of class \code{vachette_data}}

\item{lm.refine}{logical; Refine landmark x-position. Default \code{FALSE}}

\item{tol.end}{numeric; Relative tolerance to determine last x open end reference}

\item{tol.noise}{numeric; Relative tolerance for landmark determination typical curves}

\item{step.x.factor}{numeric; x-axis extension factor to search for last x, i.e., to determine where close enough to asymptote}

\item{ngrid.fit}{numeric; number of grid points in last query segment for matching last reference segment}

\item{window}{integer; size (gridpoints) of Savitzky Golay smoothing window for landmark position determination}

\item{window.d1.refine}{integer; size (gridpoints) of Savitzky Golay smoothing window for refinement of first derivative landmark position}

\item{window.d2.refine}{integer; size (gridpoints) of Savitzky Golay smoothing window for refinement of second derivative landmark position.}
}
\description{
Apply vachette transformations
}
