% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vachette_data.R
\name{apply_transformations}
\alias{apply_transformations}
\alias{apply_transformations.vachette_data}
\title{Apply vachette transformations}
\usage{
apply_transformations(vachette_data, ...)

\method{apply_transformations}{vachette_data}(
  vachette_data,
  LM_REFINE = FALSE,
  tolend = 0.001,
  tolnoise = 1e-08,
  step.x.factor = 1.5,
  ngrid.open.end = 100
)
}
\arguments{
\item{vachette_data}{Object of class \code{vachette_data}}

\item{ngrid.open.end}{}
}
\description{
Apply vachette transformations
}
