% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vachette_data.R
\name{apply_transformations}
\alias{apply_transformations}
\alias{apply_transformations.vachette_data}
\title{Apply vachette transformations}
\usage{
apply_transformations(vachette_data, ...)

\method{apply_transformations}{vachette_data}(
  vachette_data,
  LM_REFINE = FALSE,
  tolend = 0.001,
  tolnoise = 1e-08,
  step.x.factor = 1.5,
  ngrid.open.end = 100,
  w.init = 17,
  w1.refine = 7,
  w2.refine = 5
)
}
\arguments{
\item{vachette_data}{Object of class \code{vachette_data}}

\item{LM_REFINE}{Logical; Set to \code{TRUE} to refine landmark}

\item{tolend}{Numeric;}

\item{tolnoise}{Numeric;}

\item{step.x.factor}{Numeric;}

\item{ngrid.open.end}{Numeric;}

\item{w.init}{Numeric; Savitzky Golay smoothing.}

\item{w1.refine}{Numeric; Savitzky Golay smoothing, first derivative.}

\item{w2.refine}{Numeric; Savitzky Golay smoothing, second derivative.}
}
\description{
Apply vachette transformations
}
