% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vachette_functions.R
\name{get.query.x.open.end}
\alias{get.query.x.open.end}
\title{Determine best shape-matching last.x of query}
\usage{
get.query.x.open.end(
  ref,
  query,
  lm.ref,
  lm.query,
  ngrid = 100,
  scaling = "linear",
  polyorder = 9
)
}
\arguments{
\item{ref}{reference curve}

\item{query}{query curve}

\item{lm.ref}{landmarks of reference curve}

\item{lm.query}{landmarks of query curve}

\item{ngrid}{number of points in last segment}

\item{scaling}{scaling of x-axis}
}
\description{
Determine best shape-matching last.x of query
}
