% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.ref.x.open.end}
\alias{get.ref.x.open.end}
\title{get.ref.x.open.end}
\usage{
get.ref.x.open.end(x, y, lm, step.x.factor = 1.5, tol = 0.01)
}
\arguments{
\item{tol}{}
}
\value{

}
\description{
Determine last x value for open end reference curve with some defaults
}
