% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{get.x.multi.landmarks}
\alias{get.x.multi.landmarks}
\title{get.x.multi.landmarks}
\usage{
get.x.multi.landmarks(x, y, w = 17, tol = 1e-09)
}
\arguments{
\item{tol}{}
}
\value{

}
\description{
get.x.multi.landmarks
}
