% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vachette_functions.R
\name{get.y.diff.ref.query}
\alias{get.y.diff.ref.query}
\title{With constant or linearly changing scaling factor}
\usage{
get.y.diff.ref.query(
  x,
  query.x.start,
  query,
  refGrid,
  ngrid = 100,
  scaling = "linear"
)
}
\arguments{
\item{x}{last.x of query}

\item{query.x.start}{first.x of query}

\item{query}{query curve}

\item{refGrid}{reference curve}

\item{ngrid}{number of points in last segment}

\item{scaling}{scaling of x-axis}
}
\description{
With constant or linearly changing scaling factor
}
