% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vachette_functions.R
\name{multi.approx}
\alias{multi.approx}
\title{Multi approx}
\usage{
multi.approx(x, y, yout, tol = 1e-09)
}
\arguments{
\item{x}{x values}

\item{y}{y values}

\item{yout}{y value to approximate}

\item{tol}{tolerance for max and min identification}
}
\description{
Find monotonic increasing and decreasing segments
Used for both derivative and asymptote identification
}
