% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{p.add.distances}
\alias{p.add.distances}
\title{p.add.distances}
\usage{
p.add.distances(vachette_data)
}
\arguments{
\item{vachette_data}{Object of class \code{vachette_data}}
}
\value{
Object of class \code{ggplot2}
}
\description{
p.add.distances
}
