% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{p.obs.excluded}
\alias{p.obs.excluded}
\title{p.obs.excluded}
\usage{
p.obs.excluded(vachette_data)
}
\arguments{
\item{vachette_data}{Object of class \code{vachette_data}}
}
\value{
Object of class \code{ggplot2}
}
\description{
p.obs.excluded
}
