% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{p.scaled.typical.curves.landmarks}
\alias{p.scaled.typical.curves.landmarks}
\title{p.scaled.typical.curves.landmarks}
\usage{
p.scaled.typical.curves.landmarks(vachette_data)
}
\arguments{
\item{vachette_data}{Object of class\code{vachette_data}}
}
\description{
p.scaled.typical.curves.landmarks
}
