% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{p.scaling.factor}
\alias{p.scaling.factor}
\title{p.scaling.factor}
\usage{
p.scaling.factor(vachette_data)
}
\arguments{
\item{vachette_data}{Object of class\code{vachette_data}}
}
\description{
p.scaling.factor
}
