% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{p.vachette}
\alias{p.vachette}
\title{p.vachette}
\usage{
p.vachette(vachette_data, log_x = FALSE)
}
\arguments{
\item{vachette_data}{Object of class \code{vachette_data}}
}
\value{
Object of class \code{ggplot2}
}
\description{
p.vachette
}
