% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vachette_functions.R
\name{refine.x.multi.landmarks}
\alias{refine.x.multi.landmarks}
\title{Refine x multi landmarks}
\usage{
refine.x.multi.landmarks(x, y, lm, tol = 1e-09, w1 = 5, w2 = 7)
}
\arguments{
\item{x}{x values}

\item{y}{y values}

\item{lm}{landmarks}

\item{tol}{tolerance for max and min identification}

\item{w1}{window size for f1}

\item{w2}{window size for f2}
}
\description{
Refine x multi landmarks
}
