% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{summary.vachette_data}
\alias{summary.vachette_data}
\title{Summarise \code{vachette_data}}
\usage{
\method{summary}{vachette_data}(x, ...)
}
\arguments{
\item{x}{An \code{vachette_data}.}

\item{...}{Additional args.}
}
\value{
Returns \code{x} invisibly.
}
\description{
Summary generic used to return information about \code{vachette_data} object
}
