% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vachette_data.R
\name{vachette_data}
\alias{vachette_data}
\title{Initialize vachette object with required data}
\usage{
vachette_data(indivsam.obs, output.typ, indivsam.vpc = NULL, vachette.covs)
}
\arguments{
\item{indivsam.obs}{Observed \code{data.frame}}

\item{output.typ}{Typ \code{data.frame}}

\item{indivsam.vpc}{VPC \code{data.frame}}

\item{vachette.covs}{chracter vector of covariate names}
}
\value{
`vachette_data`
}
\description{
Initialize vachette object with required data
}
\examples{
indivsam.obs <- read.csv(system.file(package = "vachette", "examples", "iv-obs.csv"))
output.typ <- read.csv(system.file(package = "vachette", "examples", "iv-typ.csv"))
indivsam.vpc <- read.csv(system.file(package = "vachette", "examples", "iv-vpc.csv"))


}
