% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vachette_data.R
\name{vachette_data}
\alias{vachette_data}
\title{Initialize vachette object with required data}
\usage{
vachette_data(
  indivsam.obs,
  output.typ,
  indivsam.vpc = NULL,
  vachette.covs,
  ref.dose,
  IIV_CORR = FALSE,
  error = c("proportional", "additive"),
  model.name = NULL
)
}
\arguments{
\item{indivsam.obs}{Observed data of class \code{data.frame}}

\item{output.typ}{Typ data of class \code{data.frame}}

\item{indivsam.vpc}{VPC data of class  \code{data.frame}}

\item{vachette.covs}{Named character vector of covariate names and reference values}

\item{ref.dose}{Reference dose value in data}

\item{IIV_CORR}{Logical; if IIV corrections performed}

\item{error}{Character; options are \code{"proportional"} or \code{"additive"}}

\item{model.name}{Character; optional model name for plot output}
}
\value{
`vachette_data`
}
\description{
Initialize vachette object with required data
}
\examples{
indivsam.obs <- read.csv(system.file(package = "vachette", "examples", "iv-obs.csv"))
output.typ <- read.csv(system.file(package = "vachette", "examples", "iv-typ.csv"))
indivsam.vpc <- read.csv(system.file(package = "vachette", "examples", "iv-vpc.csv"))


}
