% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vachette_data.R
\name{vachette_data}
\alias{vachette_data}
\title{Initialize vachette object with required data}
\usage{
vachette_data(
  obs.data,
  typ.data,
  sim.data = NULL,
  covariates,
  ref.dosenr,
  iiv.correction = FALSE,
  error.model = c("proportional", "additive"),
  model.name = NULL,
  mappings = NULL
)
}
\arguments{
\item{obs.data}{data.frame; Observed data}

\item{typ.data}{data.frame; Typical (population) curves}

\item{sim.data}{data.frame; Simulated (VPC) data}

\item{covariates}{named character vector; Covariate names with reference values in vachette transformation}

\item{ref.dosenr}{integer; Number of doses given up through start of reference region}

\item{iiv.correction}{logical; Apply inter-individual variability correction. Default \code{FALSE}}

\item{error.model}{character; Applied error model, \code{"proportional"} or \code{"additive"}. Default \code{"proportional"}}

\item{model.name}{character; Optional model name for plot output}

\item{mappings}{named character vector;  Optional mappings to be included if column names in input \code{data.frame} differ from required column names}
}
\value{
\code{vachette_data}
}
\description{
Initialize vachette object with required data
}
\details{
Required column names in input \code{data.frame} are:
\itemize{
 \item{"ID"}{ - Subject ID}
 \item{"x"}{ - Typically time}
 \item{"PRED"}{ - Population prediction, required if \code{iiv.correction = TRUE}}
 \item{"IPRED"}{ - Individual prediction, required if \code{iiv.correction = TRUE}}
 \item{"OBS"}{ - DV}
}
}
\examples{
obs.data <- read.csv(system.file(package = "vachette", "examples", "iv-obs.csv"))
typ.data <- read.csv(system.file(package = "vachette", "examples", "iv-typ.csv"))
sim.data <- read.csv(system.file(package = "vachette", "examples", "iv-vpc.csv"))


}
